Refactor: 기획서 변경으로인한 코드 위치 이동
