+++ CloudNotes/CloudNotes/CloudNotes.xcdatamodeld/CloudNotes.xcdatamodel/contents
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="17709" systemVersion="20C69" minimumToolsVersion="Automatic" sourceLanguage="Swift" usedWithCloudKit="YES" userDefinedModelVersionIdentifier="">
-    <entity name="NoteItem" representedClassName="NoteItem" syncable="YES" codeGenerationType="class">
+    <entity name="NoteItem" representedClassName="NoteItem" syncable="YES">
         <attribute name="body" optional="YES" attributeType="String"/>
         <attribute name="lastModifiedDate" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="title" optional="YES" attributeType="String"/>
